REFACTOR: Merge session and auhenticatior.cookie files.
